requirements.txt
streamlit 
# utilities
numpy
pandas

# face recognition
scikit-learn
dlib-bin
# dlib-20.0.99-cp313-cp313-win_amd64.whl
git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0
#database
supabase #database
bcrypt #creating hashes for passwords

segno #qr code
pillow #preprocessing


#audio propecssing

librosa
resemblyzer